moved ScriptCsProjectSystem into a dedicated file
Conflicts:
	src/OmniSharp/Startup.cs
